line-bot-tutorial-master/requirements.txt
line-bot-sdk
flask
py_translator==1.8.9
